requirements.txt
Flask
gunicorn
Flask-MySQLdb
Flask-Uploads
Flask-WTF
WTForms
mysqlclient
python-dotenv